// src/types.ts
export enum Rank {
  HighCard = 'High Card',
  OnePair = 'One Pair',
  TwoPairs = 'Two Pairs',
  ThreeOfAKind = 'Three of a Kind',
  Straight = 'Straight',
  Flush = 'Flush',
  FullHouse = 'Full House',
  FourOfAKind = 'Four of a Kind',
  StraightFlush = 'Straight Flush',
  RoyalFlush = 'Royal Flush'
}